create separat type declaraction file
