+++ src/App.tsx
@@ -1,25 +1,7 @@
 import axios from "axios";
 import { FC, useEffect, useState } from "react";
 import User from "./components/User";
-
-interface AppProps {
-  title: string;
-}
-
-export interface Name {
-  first: string;
-  last: string;
-}
-
-export interface Login {
-  uuid: string;
-}
-
-export interface Users {
-  name: Name;
-  login: Login;
-  email: string;
-}
+import { AppProps, Users } from "./App.types";
 
 const App: FC<AppProps> = ({ title }) => {
   const [users, setUsers] = useState<Users[]>([]);
