+++ src/App.types.ts
@@ -1,0 +1,18 @@
+export interface AppProps {
+  title: string;
+}
+
+export interface Name {
+  first: string;
+  last: string;
+}
+
+export interface Login {
+  uuid: string;
+}
+
+export interface Users {
+  name: Name;
+  login: Login;
+  email: string;
+}
